Calculation of objective function values
Added functionality to calculate the objective function terms. These can
be used to do pathology detection and evaluation of segmentation
uncertainty. Also added switch priorVolumesPaper, which sets a
non-uniform prior distribution for the appearance models. Due to a bug,
we used this non-uniform distribution to obtain our results for the 3-D
dataset. With uniform prior results slightly decrease to 2.73 \mum.
For reproducibility of our results and recommended for usage with
volumetric data from Spectralis scanner.